+++ src/Ninja/G1CP/Content/Fixes/gamesave.d
@@ -35,7 +35,6 @@ func void G1CP_GamesaveFixes_Apply() {
         G1CP_133_DE_LogEntryScatty();                   // #133
         G1CP_143_DE_LogEntryBuster();                   // #143
         G1CP_193_GateSwitchesStuck();                   // #193
-        G1CP_204_LogEntryCavalornTutor();               // #204
     };
 };
 
@@ -55,6 +54,5 @@ func void G1CP_GamesaveFixes_Revert() {
         G1CP_133_DE_LogEntryScattyRevert();             // #133
         G1CP_143_DE_LogEntryBusterRevert();             // #143
         G1CP_193_GateSwitchesStuckRevert();             // #193
-        G1CP_204_LogEntryCavalornTutorRevert();         // #204
     };
 };