+++ src/Ninja/G1CP/Content/Tests/test204.d
@@ -1,36 +1,28 @@
 /*
  * #204 Cavalorn not listed as tutor
  *
- * Check if the dialog creates the topic if it did not exist and if the log entry appears when applying the fix.
+ * Check if the dialog creates the topic if it did not exist.
  *
- * Expected behavior: The log topic is correctly created in both cases.
+ * Expected behavior: The log topic is correctly created.
  */
 func int G1CP_Test_204() {
     var int passed; passed = TRUE;
 
     // Parameters of this fix
     const string topicSymbName = "GE_TeacherOW";
-    const string infoSymbName  = "DIA_Fingers_WhereCavalorn";
     const string funcSymbName  = "DIA_Fingers_WhereCavalorn_Info";
     const string tempTopicName = "G1CP Test 204";
 
-    // Check if dialog exists
-    var int infoId; infoId = MEM_GetSymbolIndex(infoSymbName);
-    if (infoId == -1) {
-        G1CP_TestsuiteErrorDetail(ConcatStrings(ConcatStrings("Dialog '", infoSymbName), "' not found"));
-        passed = FALSE;
-    };
-
     // Check if dialog function exists
-    var int funcId; funcId = MEM_GetSymbolIndex(funcSymbName);
+    var int funcId; funcId = G1CP_GetFuncID(funcSymbName, "void|none");
     if (funcId == -1) {
         G1CP_TestsuiteErrorDetail(ConcatStrings(ConcatStrings("Dialog function '", funcSymbName), "' not found"));
         passed = FALSE;
     };
 
     // Check if the constant exists
-    if (MEM_GetSymbolIndex(topicSymbName) == -1) {
-        G1CP_TestsuiteErrorDetail(ConcatStrings(ConcatStrings("Variable '", topicSymbName), "' not found"));
+    if (!G1CP_IsStringConst(topicSymbName, 0)) {
+        G1CP_TestsuiteErrorDetail(ConcatStrings(ConcatStrings("String constant '", topicSymbName), "' not found"));
         passed = FALSE;
     };
 
@@ -41,8 +33,6 @@ func int G1CP_Test_204() {
 
     // Retrieve the content of the log topic string constant
     var string topic; topic = G1CP_GetStringConst(topicSymbName, 0, "G1CP invalid string");
-
-    // First test: Check if the dialog function creates the topic if it did not exist beforehand
 
     // Rename the log topic if it already exists
     G1CP_LogRenameTopic(topic, tempTopicName);
@@ -70,35 +60,7 @@ func int G1CP_Test_204() {
         passed = FALSE;
     };
 
-    // Clean up
-    G1CP_LogRemoveTopic(topic);
-
-    // Second test: Check if the log entry is created on applying the fix
-    // This second test is specific to how the G1CP fixes the bug and will fail if a mod fixed the issue in another way
-    var int r;
-
-    // Backup the told status of the dialog
-    var int toldBak; toldBak = Npc_KnowsInfo(hero, infoId);
-
-    // Revert the fix
-    r = G1CP_204_LogEntryCavalornTutorRevert();
-
-    // Set the dialog to told
-    G1CP_SetInfoTold(infoSymbName, TRUE);
-
-    // Re-apply the fix
-    r = G1CP_204_LogEntryCavalornTutor();
-
-    // Check if the topic was created
-    if (!G1CP_LogGetTopic(topic)) {
-        G1CP_TestsuiteErrorDetail("Log topic was not auto-created on applying the fix");
-        passed = FALSE;
-    };
-
     // Revert everything
-    r = G1CP_204_LogEntryCavalornTutorRevert();
-    G1CP_SetInfoTold(infoSymbName, toldBak);
-    r = G1CP_204_LogEntryCavalornTutor();
     G1CP_LogRemoveTopic(topic);
     G1CP_LogRenameTopic(tempTopicName, topic);